+++ Dockerfile
@@ -34,6 +34,7 @@ COPY manage.py ./
 COPY docker-entrypoint.sh /usr/local/bin
 COPY dataset_retriever dataset_retriever
 COPY task_runner task_runner
+COPY oauth_connector oauth_connector
 
 RUN chmod +x /usr/local/bin/docker-entrypoint.sh
