+++ oauth_connector/urls.py
@@ -1,0 +1,10 @@
+from django.urls import include, path, re_path
+from django.conf import settings
+from .views import OAuth2LoginView, OAuth2CallbackView
+from django.http import HttpResponse
+
+urlpatterns = [
+    path('callback/', OAuth2CallbackView.as_view(), name='callback'),
+    path('test/', lambda request: HttpResponse('This is a test'), name='test'),
+    path('', OAuth2LoginView.as_view(), name='login'),
+]
