+++ task_runner/urls.py
@@ -1,8 +1,8 @@
-from django.urls import include, path, re_path
-from django.conf import settings
-from .views import OAuth2LoginView, OAuth2CallbackView
+# from django.urls import include, path, re_path
+# from django.conf import settings
+# from .views import OAuth2LoginView, OAuth2CallbackView
 
-urlpatterns = [
-    path('callback/', OAuth2CallbackView.as_view(), name='callback'),
-    path('', OAuth2LoginView.as_view(), name='login'),
-]
+# urlpatterns = [
+#     path('callback/', OAuth2CallbackView.as_view(), name='callback'),
+#     path('', OAuth2LoginView.as_view(), name='login'),
+# ]
